Split up egocentric test.
